added notifications

+++ src/model/notification_db.php
@@ -6,14 +6,19 @@ class NotificationDB {
         $db->query_prepared("INSERT INTO notification ('user_id', 'object_id', 'object_type', 'type') VALUES ($user_id, $object_id, ?, ?)", "ss", $object_type, $type);
     }
 
-    public static function remove_notification($notification_id) {
+    public static function remove_notification($user_id, $notification_id) {
         $db = SimpleDB::Singleton();
-        $db->query("DELETE FROM notification WHERE id=$notification_id");
+        $db->query("DELETE FROM notification WHERE id=$notification_id AND user_id=$user_id");
     }
 
     public static function get_notifications($user_id) {
         $db = SimpleDB::Singleton();
-        return $db->fetch("SELECT * FROM notification WHERE user_id=$user_id");
+        return $db->fetch_multiple("
+        SELECT notification.id, user.username, notification.created_datetime, notification.type, notification.object_id, notification.object_type
+        FROM notification
+        JOIN user ON user.id=notification.user_id
+        WHERE user_id=$user_id
+        ");
     }
 }
